add func to pages
