+++ client/components/Products.jsx
@@ -1,6 +1,3 @@
-'use client'
-
-import { useState, useEffect } from 'react'
 import ProductCard from './ProductCard'
 
 const ProductList = ({ data, onEdit, onDelete }) => {
@@ -24,49 +21,14 @@ const ProductList = ({ data, onEdit, onDelete }) => {
     )
 }
 
-const Products = () => {
-    const [products, setProducts] = useState([])
-
-    useEffect(() => {
-        const getProducts = async () => {
-            const res = await fetch(`http://localhost:8000/api/products/`)
-            const data = await res.json()
-
-            setProducts(data)
-        }
-
-        getProducts()
-    }, [])
-
-    const editProduct = async (id) => {
-
-        console.log('edit', id)
-    }
-
-    const deleteProduct = async (id) => {
-        const hasConfirmed = confirm("Are you sure you want to delete?")
-        if (hasConfirmed) {
-            try {
-                await fetch(`http://localhost:8000/api/products/${id}`, {
-                    method: 'DELETE',
-                }).then(() => {
-                    setProducts(products.filter((prod) => prod.id !== id))
-                }).catch(err => {
-                    console.log(err)
-                  })
-            } catch (error) {
-                console.log(error)
-            }
-        }
-        // console.log('delete', id)
-    }
-
+const Products = ({products, onEdit, onDelete}) => {
+    
     return (
         <div>
             <ProductList
                 data={products}
-                onEdit={editProduct}
-                onDelete={deleteProduct} />
+                onEdit={onEdit}
+                onDelete={onDelete} />
         </div>
 
     )
